set up routing, login, registration, and auth flow
